refactor(test): convert split_fn/tests.rs to tests/success/failure/edge pattern
