+++ crates/core/src/eval/functions/text/split_fn/tests/mod.rs
@@ -1,0 +1,3 @@
+mod success;
+mod failure;
+mod edge;
